test: give bun test the database mock vitest already had
`__test__/setup.ts` mocks `@dokploy/server/db` so that tests importing the
`@dokploy/server` barrel do not open a TCP connection to PostgreSQL on import.
vitest installs it through `setupFiles`. Bun had no equivalent - there was no
bunfig.toml at all - so the twelve directories already on `bun test` ran without
that protection.

Nothing is failing today, which is why this is easy to miss: those suites happen
not to reach the barrel, or mock the db themselves. But the guard vitest has,
bun did not, and every directory ported from here on inherits the gap.

Adds apps/dokploy/bunfig.toml with a preload. Running before any test module
loads is the whole point: `mock.module` is not hoisted, so installing this from
inside a test file would already be too late for a connection opened at import
time.

The mock shape now lives in __test__/db-mock.ts and both runners install it,
rather than being duplicated and left to drift. vitest imports it dynamically
inside the factory because `vi.mock` is hoisted above the import statements.

Verified the preload actually takes effect rather than being silently ignored:
a probe importing `dbUrl` from the barrel gets the mock value, and
`db.query.anything.findMany()` resolves through the Proxy.

  bun test    270 pass, unchanged
  vitest      595 pass, unchanged (same 4 pre-existing swarm failures)
  typecheck   0 errors, all four packages
